refactor: move ZIP handling to dedicated ziputils package
Moved the ZIP file handling code from the UI package to a dedicated ziputils package. Updated the UI and web server code to use the new package. This improves code organization and separation of concerns.
